CTCP-687 Move AddAnotherAuthorisation
