+++ conf/app.identification.authorisation.routes
@@ -1,9 +1,4 @@
 
-GET        /:mrn/identification/authorisation/add-another-authorisation                        controllers.identification.authorisation.AddAnotherAuthorisationController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = NormalMode)
-POST       /:mrn/identification/authorisation/add-another-authorisation                        controllers.identification.authorisation.AddAnotherAuthorisationController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = NormalMode)
-GET        /:mrn/identification/authorisation/change-add-another-authorisation                 controllers.identification.authorisation.AddAnotherAuthorisationController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = CheckMode)
-POST       /:mrn/identification/authorisation/change-add-another-authorisation                 controllers.identification.authorisation.AddAnotherAuthorisationController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = CheckMode)
-
 GET        /:mrn/identification/authorisation/:index/authorisation-reference-number            controllers.identification.authorisation.AuthorisationReferenceNumberController.onPageLoad(mrn: MovementReferenceNumber, index: Index, mode: Mode = NormalMode)
 POST       /:mrn/identification/authorisation/:index/authorisation-reference-number            controllers.identification.authorisation.AuthorisationReferenceNumberController.onSubmit(mrn: MovementReferenceNumber, index: Index, mode: Mode = NormalMode)
 GET        /:mrn/identification/authorisation/:index/change-authorisation-reference-number     controllers.identification.authorisation.AuthorisationReferenceNumberController.onPageLoad(mrn: MovementReferenceNumber, index: Index, mode: Mode = CheckMode)
