+++ conf/app.identification.routes
@@ -18,3 +18,10 @@ GET        /:mrn/identification/identification-number                        con
 POST       /:mrn/identification/identification-number                        controllers.identification.IdentificationNumberController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = NormalMode)
 GET        /:mrn/identification/change-identification-number                 controllers.identification.IdentificationNumberController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = CheckMode)
 POST       /:mrn/identification/change-identification-number                 controllers.identification.IdentificationNumberController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = CheckMode)
+
+GET        /:mrn/identification/authorisation/add-another-authorisation                        controllers.identification.AddAnotherAuthorisationController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = NormalMode)
+POST       /:mrn/identification/authorisation/add-another-authorisation                        controllers.identification.AddAnotherAuthorisationController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = NormalMode)
+GET        /:mrn/identification/authorisation/change-add-another-authorisation                 controllers.identification.AddAnotherAuthorisationController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = CheckMode)
+POST       /:mrn/identification/authorisation/change-add-another-authorisation                 controllers.identification.AddAnotherAuthorisationController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = CheckMode)
+
+
